refactor(viewer): extract useViewerAnnotations from GeoViewer
Decomposition 1/4, content-neutral: pin list, leave guard, arming,
editor draft, mutations, highlight/locate, and GeoJSON import move into
a hook. The measure mutual-exclusion crosses as an explicit onToggle
callback; all downstream prop names unchanged, tests untouched.
